refactor: 💡 move API config to libapi
