+++ services/api/src/api/app.py
@@ -2,6 +2,7 @@
 # Copyright 2022 The HuggingFace Authors.
 
 import uvicorn
+from libapi.config import UvicornConfig
 from libapi.jwt_token import fetch_jwt_public_key
 from libcommon.log import init_logging
 from libcommon.processing_graph import ProcessingGraph
@@ -14,7 +15,7 @@ from starlette.middleware.gzip import GZipMiddleware
 from starlette.routing import Route
 from starlette_prometheus import PrometheusMiddleware
 
-from api.config import AppConfig, EndpointConfig, UvicornConfig
+from api.config import AppConfig, EndpointConfig
 from api.routes.endpoint import EndpointsDefinition, create_endpoint
 from api.routes.healthcheck import healthcheck_endpoint
 from api.routes.metrics import create_metrics_endpoint
